+++ test/td_dd/data_structures/file_bulk_update_events_test.exs
@@ -1,51 +1,51 @@
-defmodule TdDd.DataStructures.CsvBulkUpdateEventsTest do
+defmodule TdDd.DataStructures.FileBulkUpdateEventsTest do
   use TdDd.DataCase
 
   alias Ecto.Changeset
-  alias TdDd.DataStructures.CsvBulkUpdateEvent
-  alias TdDd.DataStructures.CsvBulkUpdateEvents
+  alias TdDd.DataStructures.FileBulkUpdateEvent
+  alias TdDd.DataStructures.FileBulkUpdateEvents
 
   setup do
-    insert(:csv_bulk_update_event,
+    insert(:file_bulk_update_event,
       user_id: 1,
-      csv_hash: "hash_1",
+      hash: "hash_1",
       filename: "file_1.csv",
       inserted_at: ~U[2020-01-01 00:00:01Z]
     )
 
-    insert(:csv_bulk_update_event,
+    insert(:file_bulk_update_event,
       user_id: 1,
-      csv_hash: "hash_1",
+      hash: "hash_1",
       filename: "file_1.csv",
       inserted_at: ~U[2020-01-02 00:00:01Z],
       status: "STARTED"
     )
 
-    insert(:csv_bulk_update_event,
+    insert(:file_bulk_update_event,
       user_id: 2,
-      csv_hash: "hash_2",
+      hash: "hash_2",
       filename: "file_2.csv",
       inserted_at: ~U[2020-01-01 00:00:01Z]
     )
 
-    insert(:csv_bulk_update_event,
+    insert(:file_bulk_update_event,
       user_id: 2,
-      csv_hash: "hash_2",
+      hash: "hash_2",
       filename: "file_2.csv",
       inserted_at: DateTime.utc_now(),
       status: "STARTED"
     )
 
-    insert(:csv_bulk_update_event,
+    insert(:file_bulk_update_event,
       user_id: 2,
-      csv_hash: "hash_3",
+      hash: "hash_3",
       filename: "file_3.csv",
       inserted_at: ~U[2020-01-01 00:00:01Z]
     )
 
-    insert(:csv_bulk_update_event,
+    insert(:file_bulk_update_event,
       user_id: 2,
-      csv_hash: "hash_3",
+      hash: "hash_3",
       filename: "file_3.csv",
       inserted_at: ~U[2020-01-02 00:00:01Z],
       status: "STARTED"
@@ -56,36 +56,36 @@ defmodule TdDd.DataStructures.CsvBulkUpdateEventsTest do
 
   test "get_by_user_id" do
     assert [
-             %CsvBulkUpdateEvent{user_id: 1},
-             %CsvBulkUpdateEvent{user_id: 1}
-           ] = CsvBulkUpdateEvents.get_by_user_id(1)
+             %FileBulkUpdateEvent{user_id: 1},
+             %FileBulkUpdateEvent{user_id: 1}
+           ] = FileBulkUpdateEvents.get_by_user_id(1)
   end
 
   test "last_event_by hash gets last event by hash" do
-    assert %CsvBulkUpdateEvent{
-             csv_hash: "hash_1",
+    assert %FileBulkUpdateEvent{
+             hash: "hash_1",
              filename: "file_1.csv",
              inserted_at: ~U[2020-01-02 00:00:01.000000Z]
-           } = CsvBulkUpdateEvents.last_event_by_hash("hash_1")
+           } = FileBulkUpdateEvents.last_event_by_hash("hash_1")
   end
 
   test "last_event_by hash check_timeout inserts ALREADY_STARTED if timeout has not yet elapsed" do
-    assert %CsvBulkUpdateEvent{
-             csv_hash: "hash_2",
+    assert %FileBulkUpdateEvent{
+             hash: "hash_2",
              filename: "file_2.csv",
              status: "ALREADY_STARTED"
-           } = CsvBulkUpdateEvents.last_event_by_hash("hash_2")
+           } = FileBulkUpdateEvents.last_event_by_hash("hash_2")
   end
 
   test "last_event_by hash check_timeout inserts TIMED_OUT if timeout has already elapsed" do
-    assert %CsvBulkUpdateEvent{
-             csv_hash: "hash_3",
+    assert %FileBulkUpdateEvent{
+             hash: "hash_3",
              filename: "file_3.csv",
              status: "TIMED_OUT"
-           } = CsvBulkUpdateEvents.last_event_by_hash("hash_3")
+           } = FileBulkUpdateEvents.last_event_by_hash("hash_3")
   end
 
-  describe "CsvBulkUpdateEvents.create_event/2" do
+  describe "FileBulkUpdateEvents.create_event/2" do
     test "missing required parameter fails" do
       params = %{
         inserted_at: "2022-04-24T11:08:18.215905Z",
@@ -100,11 +100,11 @@ defmodule TdDd.DataStructures.CsvBulkUpdateEventsTest do
                :error,
                %Changeset{
                  errors: [
-                   csv_hash: {"can't be blank", [validation: :required]},
+                   hash: {"can't be blank", [validation: :required]},
                    filename: {"can't be blank", [validation: :required]}
                  ]
                }
-             } = CsvBulkUpdateEvents.create_event(params)
+             } = FileBulkUpdateEvents.create_event(params)
     end
   end
 end